[ACS-6630] Revert move of ActionRef, RuleRef and RouteRef from extensions library to js-api library
