+++ lib/extensions/src/lib/config/action.extensions.ts
@@ -47,3 +47,9 @@ export interface ContentActionRef extends ExtensionElement {
         visible?: string;
     };
 }
+
+export interface ActionRef {
+    id: string;
+    type: string;
+    payload?: string;
+}
