+++ lib/extensions/src/lib/config/routing.extensions.ts
