+++ lib/extensions/src/lib/evaluators/core.evaluators.spec.ts
@@ -16,10 +16,9 @@
  */
 
 import { every, not, some } from './core.evaluators';
-import { RuleParameter } from '@alfresco/js-api';
+import { RuleParameter } from '../config/rule.extensions';
 
 describe('Core Evaluators', () => {
-
     const context: any = {
         getEvaluator: (key: string) => {
             switch (key) {
