+++ lib/extensions/src/lib/services/extension-loader.service.ts
@@ -17,11 +17,12 @@
 
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { ContentActionRef, ContentActionType } from '../config/action.extensions';
+import { ActionRef, ContentActionRef, ContentActionType } from '../config/action.extensions';
 import { ExtensionElement } from '../config/extension-element';
 import { filterEnabled, getValue, mergeObjects, sortByOrder } from '../config/extension-utils';
 import { ExtensionConfig, ExtensionRef } from '../config/extension.config';
-import { ActionRef, RouteRef, RuleRef } from '@alfresco/js-api';
+import { RouteRef } from '../config/routing.extensions';
+import { RuleRef } from '../config/rule.extensions';
 
 @Injectable({
     providedIn: 'root'
