+++ lib/js-api/src/api/content-rest-api/model/index.ts
@@ -23,7 +23,6 @@ export * from './actionDefinitionListList';
 export * from './actionExecResult';
 export * from './actionExecResultEntry';
 export * from './actionParameterDefinition';
-export * from './actionRef';
 export * from './activity';
 export * from './activityEntry';
 export * from './activityPaging';
@@ -160,8 +159,6 @@ export * from './renditionEntry';
 export * from './renditionPaging';
 export * from './renditionPagingList';
 export * from './revertBody';
-export * from './routeRef';
-export * from './ruleRef';
 export * from './sharedLink';
 export * from './sharedLinkBodyCreate';
 export * from './sharedLinkBodyEmail';
